pyproject.toml
[project]
name = "file-big-ia"
version = "0.1.0"
description = ""
authors = [
    {name = "Gonzalo Ayuso",email = "gonzalo123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents-tools (>=0.2.16,<0.3.0)",
    "pymupdf (>=1.26.6,<2.0.0)",
    "python-calamine (>=0.3.0,<1.0.0)",
    "openpyxl (>=3.1.0,<4.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[tool.poetry]
packages = [{include = "file_big"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
